+++ src/utils/completeOptions.test.ts
@@ -90,7 +90,7 @@ describe('completeOptions', () => {
       const completed = completeOptions({ debug: true })
       const expected: CollectorOptions = {
         authKey: '',
-        requestInterval: 5000,
+        requestInterval: 1000,
         debug: true,
         maxDelay: 500,
         gravityServerUrl: GRAVITY_SERVER_ADDRESS,
@@ -112,7 +112,7 @@ describe('completeOptions', () => {
       })
       const expected: CollectorOptions = {
         authKey: '',
-        requestInterval: 5000,
+        requestInterval: 1000,
         debug: true,
         maxDelay: 15,
         gravityServerUrl: GRAVITY_SERVER_ADDRESS,
@@ -138,7 +138,7 @@ describe('completeOptions', () => {
         })
         const expected: CollectorOptions = {
           authKey: '123-456-789',
-          requestInterval: 5000,
+          requestInterval: 1000,
           debug: false,
           maxDelay: 0,
           gravityServerUrl: GRAVITY_SERVER_ADDRESS,
@@ -178,7 +178,7 @@ describe('completeOptions', () => {
         })
         const expected: CollectorOptions = {
           authKey: '123-456-789',
-          requestInterval: 5000,
+          requestInterval: 1000,
           debug: false,
           maxDelay: 0,
           gravityServerUrl: 'http://localhost:3000',
@@ -198,7 +198,7 @@ describe('completeOptions', () => {
         })
         const expected: CollectorOptions = {
           authKey: '123-456-789',
-          requestInterval: 5000,
+          requestInterval: 1000,
           debug: false,
           maxDelay: 0,
           gravityServerUrl: GRAVITY_SERVER_ADDRESS,
@@ -218,7 +218,7 @@ describe('completeOptions', () => {
         })
         const expected: CollectorOptions = {
           authKey: '123-456-789',
-          requestInterval: 5000,
+          requestInterval: 1000,
           debug: false,
           maxDelay: 0,
           gravityServerUrl: 'http://localhost:3000',